Break import cycles
