+++ src/components/views/dialogs/spotlight/PublicRoomResultDetails.tsx
@@ -9,7 +9,7 @@ Please see LICENSE files in the repository root for full details.
 import React from "react";
 import { type IPublicRoomsChunkRoom } from "matrix-js-sdk/src/matrix";
 
-import { linkifyAndSanitizeHtml } from "../../../../HtmlUtils";
+import { linkifyAndSanitizeHtml } from "../../../../Linkify";
 import { _t } from "../../../../languageHandler";
 import { getDisplayAliasForAliasSet } from "../../../../Rooms";
