+++ src/renderer/utils.tsx
@@ -7,9 +7,13 @@ Please see LICENSE files in the repository root for full details.
 
 import React, { type JSX } from "react";
 import { type DOMNode, Element, type HTMLReactParserOptions, Text } from "html-react-parser";
-import { type ParentNode } from "domhandler/lib/node";
 import { type MatrixEvent, type Room } from "matrix-js-sdk/src/matrix";
 import { type Opts } from "linkifyjs";
+
+/**
+ * The type of a parent node of an element, normally exported by domhandler but that is not a direct dependency of ours
+ */
+export type ParentNode = NonNullable<Element["parentNode"]>;
 
 /**
  * Returns the text content of a node if it is the only child and that child is a text node
